# basicprotocols/group_chat/api/swagger/docs/group_chat_swagger_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Group Chat Service API documentation, including database queries, group management, community management and other functions",
        "title": "Group Chat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:7568",
    "basePath": "/",
    "paths": {
...
        "/api/db/community/info": {
            "get": {
                "description": "Query TalkCommunityInfoCollection data based on communityId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get community information based on communityId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "communityId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/db/community/join": {
            "get": {
                "description": "Query TalkCommunityJoinCollection data based on communityId or pinId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get community join records based on communityId or pinId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "communityId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "string",
                        "description": "PinID",
                        "name": "pinId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Limit count",
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/db/community/person": {
            "get": {
                "description": "Query TalkCommunityPersonCollection data based on communityId or metaId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get community member list based on communityId or metaId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "communityId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "string",
                        "description": "MetaID",
                        "name": "metaId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Limit count",
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/db/group/info": {
            "get": {
                "description": "Query TalkGroupInfoCollection data based on groupId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get group information based on groupId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/db/group/join": {
            "get": {
                "description": "Query TalkGroupJoinCollection data based on groupId or pinId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get group join records based on groupId or pinId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "string",
                        "description": "PinID",
                        "name": "pinId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Limit count",
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/db/group/person": {
            "get": {
                "description": "Query TalkGroupPersonCollection data based on groupId or metaId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get group member list based on groupId or metaId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "string",
                        "description": "MetaID",
                        "name": "metaId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "type": "integer",
                        "description": "Limit count",
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Parameter error",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/db/stats": {
            "get": {
                "description": "Get statistics information for all database collections",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Database Query"
                ],
                "summary": "Get database statistics information",
                "responses": {
                    "200": {
                        "description": "Statistics information",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
}